Week 11 product

+++ src/main/java/com/kihong/learn/java/phoneshop/service/BrandService.java
@@ -9,8 +9,8 @@ import com.kihong.learn.java.phoneshop.entity.Brand;
 
 public interface BrandService {
    Brand Create(Brand brand);
-   Brand getById(Integer id);
-   Brand Update(Integer id,Brand brandUpdate); 
+   Brand getById(Long id);
+   Brand Update(Long id,Brand brandUpdate); 
 //   List<Brand> getBrand();
    List<Brand> getBrand(String name);
    //List<Brand> getBrand(Map<String, String> params);
